+++ exchangerates/src/main/java/com/roman/kubik/exchangerates/domain/usecase/ExchangeRatesUseCase.kt
@@ -4,9 +4,10 @@ import com.roman.kubik.core.model.Result
 import com.roman.kubik.currency.Currency
 import com.roman.kubik.exchangerates.domain.model.CurrencyRate
 import kotlinx.coroutines.flow.Flow
+import java.math.BigDecimal
 
 interface ExchangeRatesUseCase {
 
-    fun changeResponder(currency: CurrencyRate, amount: Double)
-    fun getRates(currency: Currency, amount: Double): Flow<Result<List<CurrencyRate>>>
+    fun changeResponder(currency: CurrencyRate, amount: BigDecimal)
+    fun getRates(currency: Currency, amount: BigDecimal): Flow<Result<List<CurrencyRate>>>
 }
